# internal/api/organizationsapi/oas_router_gen.go
// Code generated by ogen, DO NOT EDIT.

package organizationsapi

import (
	"net/http"
	"net/url"
	"strings"

	"github.com/ogen-go/ogen/uri"
)

var (
	rn5AllowedHeaders = map[string]string{
		"POST": "Content-Type",
	}
	rn1AllowedHeaders = map[string]string{
		"POST": "Content-Type",
	}
	rn3AllowedHeaders = map[string]string{
		"PUT": "Content-Type",
	}
	rn4AllowedHeaders = map[string]string{
		"POST": "Content-Type",
	}
	rn9AllowedHeaders = map[string]string{
		"PUT": "Content-Type",
	}
)

func (s *Server) cutPrefix(path string) (string, bool) {
	prefix := s.cfg.Prefix
	if prefix == "" {
		return path, true
	}
	if !strings.HasPrefix(path, prefix) {
		// Prefix doesn't match.
		return "", false
	}
	// Cut prefix from the path.
	return strings.TrimPrefix(path, prefix), true
}

// ServeHTTP serves http request as defined by OpenAPI v3 specification,
// calling handler that matches the path or returning not found error.
func (s *Server) ServeHTTP(w http.ResponseWriter, r *http.Request) {
	elem := r.URL.Path
	elemIsEscaped := false
	if rawPath := r.URL.RawPath; rawPath != "" {
		if normalized, ok := uri.NormalizeEscapedPath(rawPath); ok {
			elem = normalized
			elemIsEscaped = strings.ContainsRune(elem, '%')
		}
	}

	elem, ok := s.cutPrefix(elem)
	if !ok || len(elem) == 0 {
		s.notFound(w, r)
		return
	}
	args := [2]string{}

	// Static code generated router with unwrapped path search.
	switch {
	default:
		if len(elem) == 0 {
			break
		}
		switch elem[0] {
		case '/': // Prefix: "/api/organizations"

			if l := len("/api/organizations"); len(elem) >= l && elem[0:l] == "/api/organizations" {
				elem = elem[l:]
			} else {
				break
			}

			if len(elem) == 0 {
				switch r.Method {
				case "GET":
					s.handleListUserOrganizationsRequest([0]string{}, elemIsEscaped, w, r)
				case "POST":
					s.handleCreateOrganizationRequest([0]string{}, elemIsEscaped, w, r)
				default:
					s.notAllowed(w, r, notAllowedParams{
						allowedMethods: "GET,POST",
						allowedHeaders: rn5AllowedHeaders,
						acceptPost:     "application/json",
						acceptPatch:    "",
					})
				}

				return
			}
			switch elem[0] {
			case '/': // Prefix: "/"

				if l := len("/"); len(elem) >= l && elem[0:l] == "/" {
					elem = elem[l:]
				} else {
					break
				}

				if len(elem) == 0 {
					break
				}
				switch elem[0] {
				case 'i': // Prefix: "invitations/accept"
					origElem := elem
					if l := len("invitations/accept"); len(elem) >= l && elem[0:l] == "invitations/accept" {
						elem = elem[l:]
					} else {
						break
					}

					if len(elem) == 0 {
						// Leaf node.
						switch r.Method {
						case "POST":
							s.handleAcceptInvitationRequest([0]string{}, elemIsEscaped, w, r)
						default:
							s.notAllowed(w, r, notAllowedParams{
								allowedMethods: "POST",
								allowedHeaders: rn1AllowedHeaders,
								acceptPost:     "application/json",
								acceptPatch:    "",
							})
						}

						return
					}

					elem = origElem
				}
				// Param: "organizationId"
				// Match until "/"
				idx := strings.IndexByte(elem, '/')
				if idx < 0 {
					idx = len(elem)
				}
				args[0] = elem[:idx]
				elem = elem[idx:]

				if len(elem) == 0 {
					switch r.Method {
					case "DELETE":
						s.handleDeleteOrganizationRequest([1]string{
							args[0],
						}, elemIsEscaped, w, r)
					case "GET":
						s.handleGetOrganizationRequest([1]string{
							args[0],
						}, elemIsEscaped, w, r)
					case "PUT":
						s.handleUpdateOrganizationRequest([1]string{
							args[0],
						}, elemIsEscaped, w, r)
					default:
						s.notAllowed(w, r, notAllowedParams{
							allowedMethods: "DELETE,GET,PUT",
							allowedHeaders: rn3AllowedHeaders,
							acceptPost:     "",
							acceptPatch:    "",
						})
					}

					return
				}
				switch elem[0] {
				case '/': // Prefix: "/"

					if l := len("/"); len(elem) >= l && elem[0:l] == "/" {
						elem = elem[l:]
					} else {
						break
					}

					if len(elem) == 0 {
						break
					}
					switch elem[0] {
					case 'i': // Prefix: "invitations"

						if l := len("invitations"); len(elem) >= l && elem[0:l] == "invitations" {
							elem = elem[l:]
						} else {
							break
						}

						if len(elem) == 0 {
							switch r.Method {
							case "GET":
								s.handleListInvitationsRequest([1]string{
									args[0],
								}, elemIsEscaped, w, r)
							case "POST":
								s.handleCreateInvitationRequest([1]string{
									args[0],
								}, elemIsEscaped, w, r)
							default:
								s.notAllowed(w, r, notAllowedParams{
									allowedMethods: "GET,POST",
									allowedHeaders: rn4AllowedHeaders,
									acceptPost:     "application/json",
									acceptPatch:    "",
								})
							}

							return
						}
						switch elem[0] {
						case '/': // Prefix: "/"

							if l := len("/"); len(elem) >= l && elem[0:l] == "/" {
								elem = elem[l:]
							} else {
								break
							}

							// Param: "invitationId"
							// Leaf parameter, slashes are prohibited
							idx := strings.IndexByte(elem, '/')
							if idx >= 0 {
								break
							}
							args[1] = elem
							elem = ""

							if len(elem) == 0 {
								// Leaf node.
								switch r.Method {
								case "DELETE":
									s.handleRevokeInvitationRequest([2]string{
										args[0],
										args[1],
									}, elemIsEscaped, w, r)
								default:
									s.notAllowed(w, r, notAllowedParams{
										allowedMethods: "DELETE",
										allowedHeaders: nil,
										acceptPost:     "",
										acceptPatch:    "",
									})
								}

								return
							}

						}

					case 'm': // Prefix: "members"

						if l := len("members"); len(elem) >= l && elem[0:l] == "members" {
							elem = elem[l:]
						} else {
							break
						}

						if len(elem) == 0 {
							switch r.Method {
							case "GET":
								s.handleListMembersRequest([1]string{
									args[0],
								}, elemIsEscaped, w, r)
							default:
								s.notAllowed(w, r, notAllowedParams{
									allowedMethods: "GET",
									allowedHeaders: nil,
									acceptPost:     "",
									acceptPatch:    "",
								})
							}

							return
						}
						switch elem[0] {
						case '/': // Prefix: "/"

							if l := len("/"); len(elem) >= l && elem[0:l] == "/" {
								elem = elem[l:]
							} else {
								break
							}

							// Param: "userId"
							// Leaf parameter, slashes are prohibited
							idx := strings.IndexByte(elem, '/')
							if idx >= 0 {
								break
							}
							args[1] = elem
							elem = ""

							if len(elem) == 0 {
								// Leaf node.
								switch r.Method {
								case "DELETE":
									s.handleRemoveMemberRequest([2]string{
										args[0],
										args[1],
									}, elemIsEscaped, w, r)
								case "PUT":
									s.handleUpdateMemberRoleRequest([2]string{
										args[0],
										args[1],
									}, elemIsEscaped, w, r)
								default:
									s.notAllowed(w, r, notAllowedParams{
										allowedMethods: "DELETE,PUT",
										allowedHeaders: rn9AllowedHeaders,
										acceptPost:     "",
										acceptPatch:    "",
									})
								}

								return
							}

						}

					}

				}

			}

		}
	}
	s.notFound(w, r)
}

// Route is route object.
type Route struct {
	name           string
	summary        string
	operationID    string
	operationGroup string
	pathPattern    string
	count          int
	args           [2]string
}

// Name returns ogen operation name.
//
// It is guaranteed to be unique and not empty.
func (r Route) Name() string {
	return r.name
}

// Summary returns OpenAPI summary.
func (r Route) Summary() string {
	return r.summary
}

// OperationID returns OpenAPI operationId.
func (r Route) OperationID() string {
	return r.operationID
}

// OperationGroup returns the x-ogen-operation-group value.
func (r Route) OperationGroup() string {
	return r.operationGroup
}

// PathPattern returns OpenAPI path.
func (r Route) PathPattern() string {
	return r.pathPattern
}

// Args returns parsed arguments.
func (r Route) Args() []string {
	return r.args[:r.count]
}

// FindRoute finds Route for given method and path.
//
// Note: this method does not unescape path or handle reserved characters in path properly. Use FindPath instead.
func (s *Server) FindRoute(method, path string) (Route, bool) {
	return s.FindPath(method, &url.URL{Path: path})
}

// FindPath finds Route for given method and URL.
func (s *Server) FindPath(method string, u *url.URL) (r Route, _ bool) {
	var (
		elem = u.Path
		args = r.args
	)
	if rawPath := u.RawPath; rawPath != "" {
		if normalized, ok := uri.NormalizeEscapedPath(rawPath); ok {
			elem = normalized
		}
		defer func() {
			for i, arg := range r.args[:r.count] {
				if unescaped, err := url.PathUnescape(arg); err == nil {
					r.args[i] = unescaped
				}
			}
		}()
	}

	elem, ok := s.cutPrefix(elem)
	if !ok {
		return r, false
	}

	// Static code generated router with unwrapped path search.
	switch {
	default:
		if len(elem) == 0 {
			break
		}
		switch elem[0] {
		case '/': // Prefix: "/api/organizations"

			if l := len("/api/organizations"); len(elem) >= l && elem[0:l] == "/api/organizations" {
				elem = elem[l:]
			} else {
				break
			}

			if len(elem) == 0 {
				switch method {
				case "GET":
					r.name = ListUserOrganizationsOperation
					r.summary = "List organizations for the current user"
					r.operationID = "listUserOrganizations"
					r.operationGroup = ""
					r.pathPattern = "/api/organizations"
					r.args = args
					r.count = 0
					return r, true
				case "POST":
					r.name = CreateOrganizationOperation
					r.summary = "Create an organization"
					r.operationID = "createOrganization"
					r.operationGroup = ""
					r.pathPattern = "/api/organizations"
					r.args = args
					r.count = 0
					return r, true
				default:
					return
				}
			}
			switch elem[0] {
			case '/': // Prefix: "/"

				if l := len("/"); len(elem) >= l && elem[0:l] == "/" {
					elem = elem[l:]
				} else {
					break
				}

				if len(elem) == 0 {
					break
				}
				switch elem[0] {
				case 'i': // Prefix: "invitations/accept"
					origElem := elem
					if l := len("invitations/accept"); len(elem) >= l && elem[0:l] == "invitations/accept" {
						elem = elem[l:]
					} else {
						break
					}

					if len(elem) == 0 {
						// Leaf node.
						switch method {
						case "POST":
							r.name = AcceptInvitationOperation
							r.summary = "Accept an organization invitation"
							r.operationID = "acceptInvitation"
							r.operationGroup = ""
							r.pathPattern = "/api/organizations/invitations/accept"
							r.args = args
							r.count = 0
							return r, true
						default:
							return
						}
					}

					elem = origElem
				}
				// Param: "organizationId"
				// Match until "/"
				idx := strings.IndexByte(elem, '/')
				if idx < 0 {
					idx = len(elem)
				}
				args[0] = elem[:idx]
				elem = elem[idx:]

				if len(elem) == 0 {
					switch method {
					case "DELETE":
						r.name = DeleteOrganizationOperation
						r.summary = "Delete an organization"
						r.operationID = "deleteOrganization"
						r.operationGroup = ""
						r.pathPattern = "/api/organizations/{organizationId}"
						r.args = args
						r.count = 1
						return r, true
					case "GET":
						r.name = GetOrganizationOperation
						r.summary = "Get organization details"
						r.operationID = "getOrganization"
						r.operationGroup = ""
						r.pathPattern = "/api/organizations/{organizationId}"
						r.args = args
						r.count = 1
						return r, true
					case "PUT":
						r.name = UpdateOrganizationOperation
						r.summary = "Update organization details"
						r.operationID = "updateOrganization"
						r.operationGroup = ""
						r.pathPattern = "/api/organizations/{organizationId}"
						r.args = args
						r.count = 1
						return r, true
					default:
						return
					}
				}
				switch elem[0] {
				case '/': // Prefix: "/"

					if l := len("/"); len(elem) >= l && elem[0:l] == "/" {
						elem = elem[l:]
					} else {
						break
					}

					if len(elem) == 0 {
						break
					}
					switch elem[0] {
					case 'i': // Prefix: "invitations"

						if l := len("invitations"); len(elem) >= l && elem[0:l] == "invitations" {
							elem = elem[l:]
						} else {
							break
						}

						if len(elem) == 0 {
							switch method {
							case "GET":
								r.name = ListInvitationsOperation
								r.summary = "List active organization invitations"
								r.operationID = "listInvitations"
								r.operationGroup = ""
								r.pathPattern = "/api/organizations/{organizationId}/invitations"
								r.args = args
								r.count = 1
								return r, true
							case "POST":
								r.name = CreateInvitationOperation
								r.summary = "Invite a member to the organization"
								r.operationID = "createInvitation"
								r.operationGroup = ""
								r.pathPattern = "/api/organizations/{organizationId}/invitations"
								r.args = args
								r.count = 1
								return r, true
							default:
								return
							}
						}
						switch elem[0] {
						case '/': // Prefix: "/"

							if l := len("/"); len(elem) >= l && elem[0:l] == "/" {
								elem = elem[l:]
							} else {
								break
							}

							// Param: "invitationId"
							// Leaf parameter, slashes are prohibited
							idx := strings.IndexByte(elem, '/')
							if idx >= 0 {
								break
							}
							args[1] = elem
							elem = ""

							if len(elem) == 0 {
								// Leaf node.
								switch method {
								case "DELETE":
									r.name = RevokeInvitationOperation
									r.summary = "Revoke an organization invitation"
									r.operationID = "revokeInvitation"
									r.operationGroup = ""
									r.pathPattern = "/api/organizations/{organizationId}/invitations/{invitationId}"
									r.args = args
									r.count = 2
									return r, true
								default:
									return
								}
							}

						}

					case 'm': // Prefix: "members"

						if l := len("members"); len(elem) >= l && elem[0:l] == "members" {
							elem = elem[l:]
						} else {
							break
						}

						if len(elem) == 0 {
							switch method {
							case "GET":
								r.name = ListMembersOperation
								r.summary = "List organization members"
								r.operationID = "listMembers"
								r.operationGroup = ""
								r.pathPattern = "/api/organizations/{organizationId}/members"
								r.args = args
								r.count = 1
								return r, true
							default:
								return
							}
						}
						switch elem[0] {
						case '/': // Prefix: "/"

							if l := len("/"); len(elem) >= l && elem[0:l] == "/" {
								elem = elem[l:]
							} else {
								break
							}

							// Param: "userId"
							// Leaf parameter, slashes are prohibited
							idx := strings.IndexByte(elem, '/')
							if idx >= 0 {
								break
							}
							args[1] = elem
							elem = ""

							if len(elem) == 0 {
								// Leaf node.
								switch method {
								case "DELETE":
									r.name = RemoveMemberOperation
									r.summary = "Remove a member from the organization"
									r.operationID = "removeMember"
									r.operationGroup = ""
									r.pathPattern = "/api/organizations/{organizationId}/members/{userId}"
									r.args = args
									r.count = 2
									return r, true
								case "PUT":
									r.name = UpdateMemberRoleOperation
									r.summary = "Update a member's role"
									r.operationID = "updateMemberRole"
									r.operationGroup = ""
									r.pathPattern = "/api/organizations/{organizationId}/members/{userId}"
									r.args = args
									r.count = 2
									return r, true
								default:
									return
								}
							}

						}

					}

				}

			}

		}
	}
	return r, false
}